requirements.txt
# Core dependencies
torch>=1.10.0
torch-geometric>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.60.0

# Optional: For Jupyter notebooks
# jupyter>=1.0.0
# matplotlib>=3.4.0
